Sculpteo: Add paramaters and DB support
Sculpteo service now works with the database. A request for a file
upload means the file has already been uploaded to Mobius DB, and the
mobius file ID is used to fetch the file data to be then uploaded to
Sculpteo.

In addition, all requestors for quotes(or future commands) can now use
standardized mobius parameter names defined in service.py:Parameter
enum. They will be mapped to the appropriate Sculteo equivalents.